cloudwatch-exporter: improve the package
The use of underscores and dashes was inconsistent, which mean that the
system unit file did not work. So, use a dash consistently, except for
the GitHub URLs where an underscore must be used. Also add config file
templates and ensure that the JAR file has a consistent name. Add proper
--before-removal and --after-removal scripts as well.

Signed-off-by: Samuli Seppänen <samuli.seppanen@puppeteers.net>